--- DepressionDetection/templates/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Depression Detection</title>
-</head>
-<body>
-    <h2>Depression Detection System</h2>
-    <form action="/submit" method="post" enctype="multipart/form-data">
-        <label>Upload Chat File:</label>
-        <input type="file" name="chat_file" accept=".txt, .csv"><br>
-
-        <label>Upload Image (Face Emotion Detection):</label>
-        <input type="file" name="image_file" accept="image/*"><br>
-
-        <input type="submit" value="Submit">
-    </form>
-</body>
-</html>
